requirements.txt
allure-pytest==2.13.5
pytest==8.3.5
pytest-xdist==3.6.1
selenium==4.30.0
Faker==37.1.0